+++ packages/base/src/components/Inputs/Input/Input.stories.tsx
@@ -1,17 +1,27 @@
-import type { StoryFn } from '@storybook/react';
+import type { Meta, StoryFn } from '@storybook/react';
 import { useRef, useState } from 'react';
 
 import Input from './Input';
-import type { InputProps } from './Input.types';
+import { InputSizeVariants, type InputProps } from './Input.types';
 import { Stack } from '../../Stack';
 import { Button } from '../../Buttons';
 import { ComponentStage } from '../../../styles/Box';
 import { Text } from '../../Text';
 
-export default {
+const meta: Meta<typeof Input> = {
   title: 'COMPONENTS/Inputs/Input',
   component: Input,
+  argTypes: {
+    sizeVar: {
+      options: Object.values(InputSizeVariants),
+      control: { type: 'radio' },
+      description: '인풋 사이즈',
+      defaultValue: 'M',
+    },
+  },
 };
+
+export default meta;
 
 export const Playground: StoryFn<InputProps> = (args) => {
   const inputRef = useRef<HTMLInputElement>(null);
@@ -26,12 +36,20 @@ export const Playground: StoryFn<InputProps> = (args) => {
 };
 
 Playground.args = {
-  placeholder: 'input 예제에요.',
+  placeholder: 'input 예제입니다',
   disabled: false,
   onClear: () => {
     return;
   },
 };
+
+Playground.parameters = {
+  design: {
+    type: 'figma',
+    url: 'https://www.figma.com/design/KBxc4vIDtpSu2JlE4tKYIx/%5BShopl-Flow%5D-Shopl-%26-Hada-%EC%9B%B9-%EC%BB%B4%ED%8F%AC%EB%84%8C%ED%8A%B8-%EA%B3%B5%ED%86%B5%ED%99%94?node-id=407-4435&m=dev',
+  },
+};
+
 export const Password: StoryFn<InputProps> = (args) => {
   const inputRef = useRef<HTMLInputElement>(null);
 
@@ -46,7 +64,7 @@ export const Password: StoryFn<InputProps> = (args) => {
 
 Password.args = {
   type: 'password',
-  placeholder: '타입이 password인 경우에요.',
+  placeholder: '타입이 password인 경우입니다.',
   disabled: false,
   onClear: () => {
     return;
@@ -78,7 +96,7 @@ export const CheckValue: StoryFn<InputProps> = (args) => {
 };
 
 CheckValue.args = {
-  placeholder: '버튼을 눌렀을 때 input 값을 확인할 수 있어요.',
+  placeholder: '버튼을 눌렀을 때 input 값을 확인할 수 있습니다.',
   disabled: false,
   onClear: () => {
     return;
@@ -98,7 +116,7 @@ export const Error: StoryFn<InputProps> = (args) => {
 Error.args = {
   disabled: false,
   isError: true,
-  placeholder: '에러 상태에요.',
+  placeholder: '에러 상태입니다.',
 };
 export const Number: StoryFn<InputProps> = (args) => {
   const [t, sT] = useState(false);