+++ packages/base/src/components/Tabs/Tabs.tsx
@@ -1,5 +1,4 @@
 import { TabsContext } from './useTabs';
-import { LayoutGroup, AnimatePresence } from 'framer-motion';
 import { useEffect, useState } from 'react';
 import type { TabsProps } from './Tabs.types';
 import { Tab } from './Tab';
@@ -23,9 +22,7 @@ const Tabs = ({ children, initialTab, onChange }: TabsProps) => {
         setActiveTab,
       }}
     >
-      <LayoutGroup>
-        <AnimatePresence>{children}</AnimatePresence>
-      </LayoutGroup>
+      {children}
     </TabsContext.Provider>
   );
 };