+++ packages/base/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @shoplflow/base
 
+## 0.32.34
+
+### Patch Changes
+
+- [#453](https://github.com/shopl/shoplflow/pull/453) [`90cb89f94e06ed0f9b6676195f659ff047e92ab4`](https://github.com/shopl/shoplflow/commit/90cb89f94e06ed0f9b6676195f659ff047e92ab4) Thanks [@dev-bomdong](https://github.com/dev-bomdong)! - docs: Checkbox, MinusButton, Radio 문서 정리
+
+## 0.32.33
+
+### Patch Changes
+
+- [#456](https://github.com/shopl/shoplflow/pull/456) [`eea935abe5a067b05f75c003a9fc806585155625`](https://github.com/shopl/shoplflow/commit/eea935abe5a067b05f75c003a9fc806585155625) Thanks [@ayaan0209](https://github.com/ayaan0209)! - Info 탭 스타일 수정했습니다.(애니메이션 삭제 및 선 너비 수정)
+
+## 0.32.32
+
+### Patch Changes
+
+- [#450](https://github.com/shopl/shoplflow/pull/450) [`7a1578a07673c8d33d858db80af45464b1c37700`](https://github.com/shopl/shoplflow/commit/7a1578a07673c8d33d858db80af45464b1c37700) Thanks [@dev-bomdong](https://github.com/dev-bomdong)! - docs: Buttons storybook 문서 업데이트
+
+## 0.32.31
+
+### Patch Changes
+
+- [#449](https://github.com/shopl/shoplflow/pull/449) [`1bd0afd4a2b656072e55f8c7b34fa9593b02b8b8`](https://github.com/shopl/shoplflow/commit/1bd0afd4a2b656072e55f8c7b34fa9593b02b8b8) Thanks [@velo-kim](https://github.com/velo-kim)! - docs: Combobox, Inputs 문서 정리
+
+## 0.32.30
+
+### Patch Changes
+
+- [#440](https://github.com/shopl/shoplflow/pull/440) [`e1709cb8bb98049850df969444a3049b5d555cc4`](https://github.com/shopl/shoplflow/commit/e1709cb8bb98049850df969444a3049b5d555cc4) Thanks [@ayaan0209](https://github.com/ayaan0209)! - ChipButton, ChipToggle 문서 정리 했습니다.
+
+## 0.32.29
+
+### Patch Changes
+
+- [#439](https://github.com/shopl/shoplflow/pull/439) [`fa0730bbe521a5f1a66796c4738b7ff13617f401`](https://github.com/shopl/shoplflow/commit/fa0730bbe521a5f1a66796c4738b7ff13617f401) Thanks [@ayaan0209](https://github.com/ayaan0209)! - Input, InputButton, SelectInputButton, TextArea 문서를 정리했습니다.
+
+## 0.32.28
+
+### Patch Changes
+
+- [#451](https://github.com/shopl/shoplflow/pull/451) [`355c0dfabe0564efb82e6449a31a32cfaa3784fa`](https://github.com/shopl/shoplflow/commit/355c0dfabe0564efb82e6449a31a32cfaa3784fa) Thanks [@velo-kim](https://github.com/velo-kim)! - S사이즈 Button gap 변경
+
 ## 0.32.27
 
 ### Patch Changes